Refactored namespaces for sequences classes
