+++ src/String/StringSequence.php
@@ -9,10 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Eophantasy\Types\Array;
+namespace Eophantasy\Types\String;
 
 use Countable;
-use Eophantasy\Types\String\StringValue;
 use Iterator;
 
 /**
